mqtt-publisher/requirements.txt
paho-mqtt==1.6.1
pydantic==2.2.1
python-dotenv==1.0.0